fastapi-backend/requirements.txt
beautifulsoup4==4.12.3
fastapi==0.75.0
matplotlib==3.9.1.post1
nltk==3.8.1
openai==1.40.1
pandas==1.5.3
Pillow==10.4.0
python-dotenv==1.0.1
python_pptx==0.6.23
Requests==2.32.3
tiktoken==0.7.0
transformers==4.43.3
typing_extensions==4.12.2
uvicorn==0.30.5
